chore: rename room for better clarity
