+++ frontend/components/CollabRoom/ChatPanel.tsx
