+++ frontend/components/CollabRoom/Resizer.tsx
