+++ frontend/pages/room.tsx
@@ -1,5 +1,5 @@
 import PageWrapper from 'components/PageWrapper';
-import Room from 'components/Room';
+import Room from 'components/CollabRoom';
 import type { NextPage } from 'next';
 
 const PeerPrepRoom: NextPage = () => {
